Updated Catalog to be an Int for queries, made example sheet
Was playing with the types of queries we'd need to generate with a Student Question, realized we need to be able to search a range of Catalogs so that's an Int now.  Made a reference sheet for some common queries.  ***Still need to add Grade to student-classes import so that we can tell students' grades, which also determines if they're currently enrolled in a given course***