Started primitive WHERE clause builder
WHERE clause structure and form set up, doesnt take data from Jerad's object yet.  Is ready to make a query based on all of the different types of data it might receive though, except for Minimum Grade.  Grade is not in StudentClass table yet so i'm gonna change the database and import after this commit