Fixed an OR bug,  fixed deleting AND button
AND sections can now be removed and added with 100% BUT more work needs done if we want it to always use the bottom button to add any div like the original program.  Ours instead needs to use the add button above the " - " button if we want to re-add sections.   Fixed an OR bug I think, the first delete OR button always deletes childeNode[0] if clicked and only when there are 2 children or more.  Still has bugs and im still looking at it, but i need to commit this before i lose it