QUERY BUILDER now handle's 200's, 300's queries
Query Builder now checks to see if user selected something like 100s, 200s.
Depending on what they chose, it makes an AND subquery for catalog that will look like one of these two:
    AND CATALOG = 244
    AND CATALOG LIKE 2%
The LIKE operator means there can be any number of any characters after the 2, so in our case, LIKE 2% means all 200's level courses
Also, I changed the SQL script on my previous commit.  The Catalog columns are now VarChars instead of ints, for 402G, as well as our LIKE clause.  We can no longer do mathematical comparisons on that column, you will need to use LIKE from now on.
So we can query ANY, PASSED, 200s, or any specific chosen value.  All that is left is to implement the hardest part:
We still don't actually check completed, taking, or scheduled for, as those are defined as having a term that is Lower, Equal To, or Greater than the current term.  That is what I will work on next.