QUERY BUILDER is done
Terms are now factored in for query builder.  You can now make any question regardless of AND's and OR's and I believe it will build the correct query, and return the correct results.  Taking/Completed/Scheduled for etc are implemented as a SWITCH statement.  An array of all category items set is made, and the indexes of those items matches 1 to 1 with the subject-cat-grade AND indexes we find.  We then switch on the value of the category data we received (cat0 = Taking, etc) matched to the AND index of a class item we find (represented by a "sub01 etc" substring(3,1) found in $stdq->data.  So if our first AND is a class option like taking, completed etc we will have cat0 = Taking and sub01, cor01, gra01 will be matched to it, so that our AND block's subquery will end with AND TERM = $currentTerm.  Examples of queries and subqueries built are still printed to the top of the screen for debugging, but I think we're all out of bugs to debug