USER CONTEXT now implemented
Student Question has a new body.onload that passes the currently signed-in user into getSubjectsForUser, which uses Ajax to do exactly what it sounds like.  The SQL query is a little weird to look at, it's available in model/model.php if you want to view it.  Basically body.onload->JS->ajax/JSON page->model ->return return return return
Also applied the same logic to loading programs, we don't get all the programs for the dropdowns anymore, just the ones associated with that user.
So basically once we have all the programs for a user and all the subjects for those programs, the dropdowns fill with all of that relevant data rather than just... all of our data.
One quirk that showed up is that we currently have subjects in our DB that we got from student-classes instead of classes, meaning subjects other than CIS and DA.  Subjects like basket weaving and native american folk dancing and shit.  So if you decide to associate those subjects with a certain program, then that program with whatever user you're testing, dont be surprised when the catalog dropdown only shows "100s, 200s, etc" because it's not going to find any matching classes in our Courses table.  BUT it will in fact be able to find students who took those classes if you search, say, CHIN ANY PASSED.  You'll find student who passed any Chinese class.  For this to work, I had to add a check on the JSON object that carries all relevant subjects to make sure it's not undefined.  The reason for this is that we really don't use the same list of subjects for both functions.  This is an important distinction!  The subjects that show up in the subject dropdown are whatever subjects the admin decided to associate with the logged in user's relevant programs.  The subjects that actually have the ability to fill catalog dropdowns are just the ones that show up in our Courses files/tables.  FOR EXAMPLE our data has CIS and DA courses in the Courses file.  BUT it has math, spanish, basket weaving, etc in Students-Classes because its all of the classes taken by any student in our major.  SO if you query on basket weaving you'll get zero catalogs to search by.  If you query on CIS or whatever, you'll get actual catalogs to choose from.
TL; DR:
User context is done