*IMPORTANT* Javascript Changes
A little background:
On VCISPROD, our rebuild was not working.  When rebuilding subjects, nothing but CIS and DA would load correctly.  For some reason, we could only rebuild subjects that already existed in the options of the Select, as CIS and DA were still hardcoded into MakeDivVisibleOr() from like Day 1.
So, we have a few changes.  For one, we are NO LONGER LISTENING FOR the event: "DOMContentLoaded".  We are now waiting until all of our ajax calls have recieved a complete response, THEN running the code that we *used to* run upon loading the DOMContent.  So, I took that code (adding listeners to saveQuestion and dropdown0 for loading a saved search) and put it into a new function, GetSerialsReady().  This function does everything we used to do on DOMContentLoaded but instead now executes after our AJAX calls are done.  The way we do this is, whenever we get an xhttpresponse from an AJAX call, we call a function: incrementAjaxCompletionCounter(), which increments a variable called AjaxCompletedCount, then if the count is equal to our constant AJAX_CALL_TOTAL, we execute GetSerialsReady().
So, we now wait the optimal time for loading serials no matter how long the ajax calls take (its not very long on vcisprod).
*ALSO IMPORTANT*
Another change was made to the hellish MakeDivVisibleOr() function.  We no longer hardcode in CIS and DA, we now take the AJAX response that gives us the Subjects associated with the logged in User and create a big string of options and values using those subjects.  Then, we add that string to the MakeDivVisibleOr() section that previously hardcoded in CIS and DA.  Now, it loads in whatever subjects they'll end up seeing on loading.
TL; DR:
Form Rebuild works on VCISPROD now