Security: AddUser
Adding a new User now automatically makes them a Reader.  This is helping with our inability to assign roles to users that we have yet to fix.  We can assign functions to Reader though, so it is not difficult to make a new user who can use our pages without being an admin.