Security Stuff
Now, if you're username is not jodonnell or admin, you get redirected to the StudentQuestion page upon logging in.
Also, removed a listener that I think went all the way back to the 370 security framework and is not needed.
Still trying to load chance.js, might remove it if I can verify that it is not needed as we can't load it over http anyway, as vcisprod wants us to use https