Back Button for Results Page
From results page, you can now click "Back" to return to your search.  On Chrome, it seems like this will maintain the state of your search (as in, all of the dropdowns will be there).  I'm not sure about other browsers.  Also, the results and student history are blue now