Example Course Question Results
Results page for Course Question is done.
Now, we just need to be able to build the query that actually gets these results.
An example query is available in SQL EXAMPLE SCRIPTS file at the bottom, it will be the header of our query builder.  The only additions to the HEADER will be the "WHERE" clause that decides specifically which classes will be returned.