Course Question Example Header
Updated the Example Scripts file to have an example of what a Course Question might look like.  This is the header for our course questions, it will return all the information we need for ALL courses.  We will need the query builder to form a WHERE clause that gets us the correct answers